+++ Changelog.md
@@ -16,21 +16,39 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - Projects are created from platforms, not templates
  - Improved Teams pages
  - Improved access permissions and descriptions
+ - Fix for content component's 'unsaved changes' error message
+
+## 4.0.16
+### Changed
+ - Fix for content component's 'unsaved changes' error message
+ - Upgraded ng-formio to 2.19.5
+ - Upgraded ng-formio-builder to 2.19.3
+
+### Removed
+ - Remove form component until recursion issue fixed.
+
+## 4.0.15
+### Fixed
+ - Some empty components were displaying incorrectly in submission grid.
+
+### Changed
+ - Upgraded ng-formio to 2.19.2
+ - Upgraded ng-formio-builder to 2.19.2
+
+## 4.0.14
+### Changed
+ - Upgraded ng-formio to 2.19.1
+ - Upgraded ng-formio-builder to 2.19.1
 
 ### Fixed
+ - All spinners to give feedback.
  - Bugs with Teams and Access
+ - Some empty components were displaying incorrectly in submission grid.
 
 ### Removed
  - Tour
  - Environment Switcher
  - Project Templates
-
-### Fixed
- - Some empty components were displaying incorrectly in submission grid.
-
-### Changed
- - Upgraded ng-formio to 2.19.1
- - Upgraded ng-formio-builder to 2.19.1
 
 ## 4.0.13
 ### Changed